requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
faiss-cpu
numpy
pandas
openai
pdfplumber
python-docx
tqdm
